+++ Sources/Sockets/TCP/TCPInternetSocket.swift
@@ -1,8 +1,8 @@
 import libc
 
-public typealias TCPDuplexProgram = TCPServer & TCPClient
+//public typealias TCPDuplexProgram = TCPServer & TCPClient
 
-public final class TCPInternetSocket: InternetSocket, TCPDuplexSocket, TCPDuplexProgram,  DuplexProgramStream {
+public final class TCPInternetSocket: InternetSocket, TCPDuplexSocket,  DuplexProgramStream {
 
     // stream
     public let hostname: String
@@ -57,6 +57,48 @@ public final class TCPInternetSocket: InternetSocket, TCPDuplexSocket, TCPDuplex
         port = resolved.port
         self.securityLayer = securityLayer
         self.isClosed = false
+    }
+
+    // MARK: Client
+
+    public func connect() throws {
+        if isClosed { throw SocketsError(.socketIsClosed) }
+        let res = libc.connect(descriptor.raw, address.raw, address.rawLen)
+        guard res > -1 else { throw SocketsError(.connectFailed) }
+    }
+
+    // MARK: Server
+
+    public func listen(queueLimit: Int32 = 4096) throws {
+        if isClosed { throw SocketsError(.socketIsClosed) }
+        let res = libc.listen(descriptor.raw, queueLimit)
+        guard res > -1 else { throw SocketsError(.listenFailed) }
+    }
+
+    public func accept() throws -> TCPInternetSocket {
+        if isClosed { throw SocketsError(.socketIsClosed) }
+        var length = socklen_t(MemoryLayout<sockaddr_storage>.size)
+        let addr = UnsafeMutablePointer<sockaddr_storage>.allocate(capacity: 1)
+        let addrSockAddr = UnsafeMutablePointer<sockaddr>(OpaquePointer(addr))
+        let clientSocketDescriptor = libc.accept(descriptor.raw, addrSockAddr, &length)
+
+        guard clientSocketDescriptor > -1 else {
+            addr.deallocate(capacity: 1)
+            if errno == SocketsError.interruptedSystemCall {
+                return try accept()
+            }
+            throw SocketsError(.acceptFailed)
+        }
+
+        let clientAddress = ResolvedInternetAddress(raw: addr)
+        let clientSocket = try TCPInternetSocket(
+            Descriptor(clientSocketDescriptor),
+            config,
+            clientAddress,
+            self.securityLayer
+        )
+
+        return clientSocket
     }
 
     // MARK: Close
