+++ Sources/Transport/Streams/ProgramStream/ServerStream.swift
@@ -1,3 +1,3 @@
 public protocol ServerStream: ProgramStream {
-    func accept() throws -> DuplexStream
+    func accept() throws -> Self
 }
